+++ src/base/ProviderInterface.php
@@ -28,9 +28,9 @@ interface ProviderInterface
 
     /**
      * Gets a concrete League provider instance
-     * @return AbstractProvider
+     * @return ProviderInterface
      */
-    public function getConfiguredProvider();
+    public function getConfiguredProvider(): ProviderInterface;
 
     /**
      * Get the URL used to authorize the token
@@ -41,12 +41,12 @@ interface ProviderInterface
 
     /**
      * Get the session security state
-     * @return string|null
+     * @return string
      */
     public function getState(): string;
 
     /**
-     * Gets an access token, returing a League access token
+     * Gets an access token, returning a League access token
      * @param $grant
      * @param array $options
      * @return AccessTokenInterface