+++ src/variables/OAuthVariable.php
@@ -18,7 +18,7 @@ class OAuthVariable
      * @param $handle
      * @return App|null
      */
-    public function getAppByHandle($handle)
+    public function getAppByHandle($handle): ?App
     {
         return Plugin::getInstance()->apps->getAppByHandle($handle);
     }